+++ src/test/03.comment.test.ts
@@ -7,13 +7,15 @@ import {
     NCFolder,
 } from "../ncClient";
 
+const instanceName = "test";
+
 // tslint:disable-next-line:only-arrow-functions
 describe("NEXCLOUD-NODE-CLIENT-COMMENT", function() {
     this.timeout(1 * 60 * 1000);
 
     it("1 add comment to file", async () => {
 
-        const client = await NCClient.clientFactory();
+        const client: NCClient = new NCClient(instanceName);
 
         let errorOccurred;
         const fileName = "/test/comments/fileComments.txt";
@@ -52,7 +54,7 @@ describe("NEXCLOUD-NODE-CLIENT-COMMENT", function() {
 
     it("2 add comment to folder", async () => {
 
-        const client = await NCClient.clientFactory();
+        const client: NCClient = new NCClient(instanceName);
 
         let errorOccurred;
         const folderName = "/test/folder/comments";
@@ -91,7 +93,7 @@ describe("NEXCLOUD-NODE-CLIENT-COMMENT", function() {
 
     it("99 delete directory", async () => {
 
-        const client = await NCClient.clientFactory();
+        const client: NCClient = new NCClient(instanceName);
 
         const dirName = "/test";